back/requirements.txt
Django==5.0.3
channels==4.1.0
channels-redis==4.1.0 # For using channels with redis
celery==5.2.7
django-celery-results # Task and state admin displays for celery
django-celery-beat # For automatic and timed tasks
djangorestframework==3.14.0
uvicorn[standard]
# [sidecar] only used to not need static files from CDN
drf-spectacular[sidecar] # renders swagger, route will only be available for admins if in production
whitenoise # Serving static files in development
django-webpack-loader # Load the static file bundles
django-cors-headers
django-spaghetti-and-meatballs # shows a sweet db render at /db
django-ipware # Best way I found to determine and track IP adresses
django-jazzmin # Waaay better admin UI
ipython # With this we can use the IPython shell on manage.py shell
django-model-utils # this package is required for our chat backend!
django-appconf # Required for django-cookie-consent
django-multiselectfield @ git+https://github.com/tbscode/django-multiselectfield@master
# I maintain a fork, this is a much used django package but the maintainer doesn't seem to be active & does't merge pr's so I forked...
django-phonenumber-field[phonenumbers]
Babel # Required for the widged of the phone number field
Pillow # Required for django image field
django-rest-passwordreset
requests # for tests mainly
twilio # For the video calls
pytest # Are used for the tests in cookie_banner
psycopg2-binary # For aws postgres sql
django-storages # For manageing s2 bucket in combinaton with boto3
boto3 # for connection loading and updloading to amazon s3
martor # Markdown editar, also used for editable tables in admin view
networkx # For scoring calculation using graphs
pytablewriter # For generating goodlooking markdown tables
Collectfast # Faster static file collection
django-hijack # Can be used by admin to login as another user
single-beat
djangorestframework-dataclasses
polib
sentry-sdk[django]
# django-minio-storage
django-revproxy
slack_sdk
openai==1.8.0
pgeocode
django-filter
bs4
livekit-api
google-cloud-translate==3.15.3
django-stubs[compatible-mypy]
